backend/requirements.txt
asgiref==3.4.0
Django==3.2.4
djangorestframework==3.12.4
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
PyJWT==2.1.0
django-cors-headers==3.7.0
faker==8.8.2
Pillow==8.2.0
